e2e tests: do not cast string[] elements to i32
A string-to-i32 cast is a legal data-pointer hop, not a value conversion, so
the checker accepts `(got[i] as i32) != (want[i] as i32)` on a string[] and
the comparison silently compares two pointers -- always unequal. That is what
CI's wit-lift self-test caught as "failed at check 1".

The previous commit's verification could not see this: it compared error
counts before and after, and this rewrite introduces no error. Only a
behavioural test catches it.

+++ internal/e2eselfhost/self_host_wit_decode_test.go
@@ -343,7 +343,7 @@ function main(): i32 {
     if (got.len() != want.len()) { return 99; }
     var i: i32 = 0;
     while (i < want.len()) {
-        if ((got[i] as i32) != (want[i] as i32)) { return i + 1; }
+        if (got[i] != want[i]) { return i + 1; }
         i = i + 1;
     }
     return 0;
